requirements.txt
opencv-python
fer
matplotlib
moviepy
pandas
numpy
seaborn
SpeechRecognition
ffmpeg-python
nltk
